Move match_suffix from misc1.c
