+++ src/misc1.h
@@ -67,7 +67,6 @@ void prepare_to_exit(void);
 void preserve_exit(void);
 void line_breakcheck(void);
 void fast_breakcheck(void);
-int match_suffix(char_u *fname);
 char_u *get_cmd_output(char_u *cmd, char_u *infile, int flags);
 void FreeWild(int count, char_u **files);
 int goto_im(void);
